requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
psycopg2==2.8.6
python-utils==2.4.0
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.25.9
azure-core==1.9.0
azure-storage-blob==12.6.0
alembic==1.4.3
lxml==4.6.2